Reorganize the lib folder files structure
