+++ lib/kind/core.rb
@@ -1,8 +1,15 @@
 # frozen_string_literal: true
 
-module Kind
-  module Core
-    require 'kind/core/kind'
-    require 'kind/core/undefined'
-  end
-end
+require 'kind/core/utils/kind'
+require 'kind/core/utils/undefined'
+
+require 'kind/core/error'
+require 'kind/core/empty'
+require 'kind/core/dig'
+require 'kind/core/try'
+require 'kind/core/presence'
+require 'kind/core/undefined'
+require 'kind/core/maybe'
+
+require 'kind/core/type_checker'
+require 'kind/core/type_checkers'
