+++ lib/kind/core/dig.rb
