+++ lib/kind/core/empty.rb
