+++ lib/kind/core/empty/constant.rb
