+++ lib/kind/core/error.rb
