+++ lib/kind/core/maybe/result.rb
