+++ lib/kind/core/maybe/typed.rb
