+++ lib/kind/core/maybe/wrappable.rb
