+++ lib/kind/core/presence.rb
