+++ lib/kind/core/try.rb
