+++ lib/kind/core/type_checkers.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+require 'kind/core/type_checkers/ruby/array'
+require 'kind/core/type_checkers/ruby/class'
+require 'kind/core/type_checkers/ruby/comparable'
+require 'kind/core/type_checkers/ruby/enumerable'
+require 'kind/core/type_checkers/ruby/enumerator'
+require 'kind/core/type_checkers/ruby/file'
+require 'kind/core/type_checkers/ruby/float'
+require 'kind/core/type_checkers/ruby/hash'
+require 'kind/core/type_checkers/ruby/integer'
+require 'kind/core/type_checkers/ruby/io'
+require 'kind/core/type_checkers/ruby/module'
+require 'kind/core/type_checkers/ruby/method'
+require 'kind/core/type_checkers/ruby/numeric'
+require 'kind/core/type_checkers/ruby/proc'
+require 'kind/core/type_checkers/ruby/queue'
+require 'kind/core/type_checkers/ruby/range'
+require 'kind/core/type_checkers/ruby/regexp'
+require 'kind/core/type_checkers/ruby/string'
+require 'kind/core/type_checkers/ruby/struct'
+require 'kind/core/type_checkers/ruby/symbol'
+require 'kind/core/type_checkers/ruby/time'
+
+require 'kind/core/type_checkers/stdlib/open_struct'
+require 'kind/core/type_checkers/stdlib/set'
+
+require 'kind/core/type_checkers/custom/boolean'
+require 'kind/core/type_checkers/custom/callable'
+require 'kind/core/type_checkers/custom/lambda'
