+++ lib/kind/core/type_checkers/custom/boolean.rb
