+++ lib/kind/core/type_checkers/custom/callable.rb
