+++ lib/kind/core/type_checkers/custom/lambda.rb
