+++ lib/kind/core/type_checkers/ruby/array.rb
