+++ lib/kind/core/type_checkers/ruby/class.rb
