+++ lib/kind/core/type_checkers/ruby/comparable.rb
