+++ lib/kind/core/type_checkers/ruby/enumerable.rb
