+++ lib/kind/core/type_checkers/ruby/enumerator.rb
