+++ lib/kind/core/type_checkers/ruby/file.rb
