+++ lib/kind/core/type_checkers/ruby/float.rb
