+++ lib/kind/core/type_checkers/ruby/hash.rb
