+++ lib/kind/core/type_checkers/ruby/integer.rb
