+++ lib/kind/core/type_checkers/ruby/io.rb
