+++ lib/kind/core/type_checkers/ruby/method.rb
