+++ lib/kind/core/type_checkers/ruby/module.rb
