+++ lib/kind/core/type_checkers/ruby/numeric.rb
