+++ lib/kind/core/type_checkers/ruby/proc.rb
