+++ lib/kind/core/type_checkers/ruby/queue.rb
