+++ lib/kind/core/type_checkers/ruby/range.rb
