+++ lib/kind/core/type_checkers/ruby/regexp.rb
