+++ lib/kind/core/type_checkers/ruby/string.rb
