+++ lib/kind/core/type_checkers/ruby/struct.rb
