+++ lib/kind/core/type_checkers/ruby/symbol.rb
