+++ lib/kind/core/type_checkers/ruby/time.rb
