+++ lib/kind/core/type_checkers/stdlib/open_struct.rb
