+++ lib/kind/core/type_checkers/stdlib/set.rb
