+++ lib/kind/core/utils/kind.rb
