+++ lib/kind/core/utils/undefined.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Kind
+  UNDEFINED = Object.new.freeze
+
+  private_constant :UNDEFINED
+end
