--- lib/kind/type_checkers.rb
@@ -1,30 +1,0 @@
-# frozen_string_literal: true
-
-require 'kind/type_checkers/core/array'
-require 'kind/type_checkers/core/class'
-require 'kind/type_checkers/core/comparable'
-require 'kind/type_checkers/core/enumerable'
-require 'kind/type_checkers/core/enumerator'
-require 'kind/type_checkers/core/file'
-require 'kind/type_checkers/core/float'
-require 'kind/type_checkers/core/hash'
-require 'kind/type_checkers/core/integer'
-require 'kind/type_checkers/core/io'
-require 'kind/type_checkers/core/module'
-require 'kind/type_checkers/core/method'
-require 'kind/type_checkers/core/numeric'
-require 'kind/type_checkers/core/proc'
-require 'kind/type_checkers/core/queue'
-require 'kind/type_checkers/core/range'
-require 'kind/type_checkers/core/regexp'
-require 'kind/type_checkers/core/string'
-require 'kind/type_checkers/core/struct'
-require 'kind/type_checkers/core/symbol'
-require 'kind/type_checkers/core/time'
-
-require 'kind/type_checkers/stdlib/open_struct'
-require 'kind/type_checkers/stdlib/set'
-
-require 'kind/type_checkers/custom/boolean'
-require 'kind/type_checkers/custom/callable'
-require 'kind/type_checkers/custom/lambda'
